% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dj94functions.R
\name{bumps}
\alias{bumps}
\alias{bumps_dj94}
\title{Bumps function}
\usage{
bumps(x, t, h, w)

bumps_dj94(x)
}
\arguments{
\item{x}{numeric. See description.}

\item{t}{numeric. See description.}

\item{h}{numeric. See description.}

\item{w}{numeric. See description.}
}
\description{
The Bumps function from Donoho and Johnstone (1994) defined as,
\deqn{
f(x) = \sum h_j K((x - t_j) / w_j), K(x) = (1 + |x|)^{-4}
}{
f(x) = sum( h_j K((x - t_j) / w_j), K(x) = (1 + |x|)^{-4})
}
The function \code{bumps} is the general function.
The function \code{bumps_dj94} uses the specific parameters in the paper,
\deqn{
(t_j) = (0.1, 0.13, 0.15, 0.23, 0.25, 0.40, 0.44, 0.65, 0.76, 0.78, 0.81)
}
and
\deqn{
(h_j) = (4, 5, 3, 4, 5, 4.2, 2.1, 4.3, 3.1, 5.1, 4.2)
}
and
\deqn{
(w_j) = (0.005, 0.005, 0.006, 0.01, 0.01, 0.03, 0.01, 0.01, 0.005, 0.008, 0.005)
}
for the range (0, 1).
}
\examples{
x <- seq(0, 1, length.out = 2048)
plot(x, bumps_dj94(x), type = "l")
}
\references{
Donoho, David L, and Johnstone, Iain M. (1994) \dQuote{Ideal spatial adaptation by wavelet shrinkage.}
\emph{Biometrika} \strong{81.3}, 425-55.
}
\seealso{
Other DonohoJohnstone1994: \code{\link{doppler}},
  \code{\link{doppler_dj94}}; \code{\link{heavisine}},
  \code{\link{heavisine_dj94}}
}

