% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BushApproval}
\alias{BushApproval}
\title{Approval Ratings for President George W. Bush}
\format{A \code{data.frame} object with \Sexpr{ncol(bsdlm::BushApproval)} variables and \Sexpr{nrow(bsdlm::BushApproval)}  observations.
\itemize{
\item start_date. Start date of the survey.
\item end_date. End date of the survey.
\item approve. Percent which approve of the president.
\item disapprove. Percent which disapprove of the president.
\item undecided. Percent undecided about the president.
}}
\usage{
BushApproval
}
\description{
Approval ratings for George W. Bush.
}
\keyword{datasets}

