% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dj94functions.R
\name{doppler}
\alias{doppler}
\alias{doppler_dj94}
\title{Doppler function}
\usage{
doppler(x, eps)

doppler_dj94(x)
}
\arguments{
\item{x}{numeric. See description.}

\item{eps}{numeric. See description.}
}
\description{
The Doppler function from Donoho and Johnstone (1994),
\deqn{
f(x) = (x (1 - x)^{\frac{1}{2}} \sin (2 \pi (1 + \epsilon) / (x + \epsilon))
}{
f(x) = (x (1 - x))^(1 / 2) * sin (2 * pi * (1 + e) / (x + e))
}
The function \code{doppler} is the general case.
The function \code{dollper_dj94} is the special case considered in the paper, \eqn{epsilon = 0.05}.
}
\examples{
x <- seq(0, 1, length.out = 2048)
plot(x, doppler_dj94(x), type = "l")
}
\references{
Donoho, David L, and Johnstone, Iain M. (1994) "Ideal spatial adaptation by wavelet shrinkage."
\emph{Biometrika} \strong{81.3}, 425-55.
}
\seealso{
Other DonohoJohnstone1994: \code{\link{bumps}},
  \code{\link{bumps_dj94}}; \code{\link{heavisine}},
  \code{\link{heavisine_dj94}}
}

