% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rbernoulli.R
\name{rbernoulli}
\alias{rbernoulli}
\title{Bernoulli distribution}
\usage{
rbernoulli(n, prob)
}
\arguments{
\item{n}{number of observations}

\item{prob}{probability of success on each trial.}
}
\value{
\code{numeric} vector of samples from the distribution.
}
\description{
Random generation for the bernoulli distribution.
}

